Updated custom quest ids to be within short range so that packets work
Second job had custom quest IDs which started from 100,000, but the packet sender needs to send a short, which has the max value of 32,767. I changed the custom IDs to start at 32,000 instead which should allow for a buffer if required in future. Bowman was missing the final 2nd job custom quest set as completed, so that's been added too.